PID adjust

+++ include/bldc.h
@@ -11,7 +11,7 @@
 #define TICKS_PER_PHASE (BLDC_TICKS / (BLDC_POLES/2) / 6)
 #define PHASES_PER_TICK (1/TICKS_PER_PHASE)
 
-void BLDC_Commute(int *current_pointer, int phase_order, int direction, int PWM);
+void BLDC_Commute(int *current_pointer, int phase_order, int direction, Uint16 PWM);
 void Init_Motor_Drvs(void);
 
 #endif // end of F28069_BLDC_H definition
